fix doc numbers for project prefix

+++ src/main/java/com/example/store/model/entities/documents/Document.java
@@ -25,7 +25,7 @@ public class Document {
     @SequenceGenerator(name = "docSequence", sequenceName = "DOC_SEQ")
     private int id;
 
-    private int number;
+    private long number;
 
     private LocalDateTime dateTime;
